消息task修复

+++ manage/src/main/java/com/smart/agriculture/manage/entity/MachineSensorEntity.java
@@ -21,7 +21,7 @@ public class MachineSensorEntity implements Serializable {
     /**
      * 条件数值
      */
-    private Integer conditionNumber;
+    private Double conditionNumber;
     /**
      * 记录时间
      */
